ID should be biginteger in resource routing table
1. What is the problem?
Currently the id in resource routing table is integer, the scope of integer is
very limited, so it should be replaced with biginteger.

2. What is the solution to the problem?
Replacing the integer with biginteger will fix this problem. What's more,
API documentation should be updated at the same time.

3. What the features need to be implemented to the Tricircle to realize
the solution?
None.

Closes-Bug: #1647513

+++ tricircle/db/migrate_repo/versions/002_resource.py
@@ -241,7 +241,7 @@ def upgrade(migrate_engine):
 
     cascaded_pods_resource_routing = sql.Table(
         'cascaded_pods_resource_routing', meta,
-        sql.Column('id', sql.Integer, primary_key=True),
+        sql.Column('id', sql.BigInteger, primary_key=True),
         sql.Column('top_id', sql.String(length=127), nullable=False),
         sql.Column('bottom_id', sql.String(length=36)),
         sql.Column('pod_id', sql.String(length=64), nullable=False),
